Move edit functionality
